Add the `repository_metadata` column to the `Node` database models
This new JSONB column will be used to store the metadata of the contents
of the file repository. It essentially is a mapping of the file object
hierarchy, where each object contains the relative filepath of it within
the repository, which is purely symbolic, and a hashkey, which is the
actual identifier that uniquely identifies the object in the repository
backend.

The column is made nullable and there is no default set. This is to make
sure to not add unnecessary bytes to the database for all the nodes that
contain no files whatsoever, which will be a decent chunk. The front-end
ORM class `Node` will return an empty dictionary if the database model
field is null, such that the clients don't have to deal with `None`.